static casted response to unsigned short

+++ mtp/ptp/Response.cpp
@@ -81,7 +81,7 @@ namespace mtp
 		std::string FormatMessage(ResponseType r)
 		{
 			char buf[1024];
-			snprintf(buf, sizeof(buf), "invalid response code %s (0x%04hx)", GetResponseName(r).c_str(), (unsigned)r);
+			snprintf(buf, sizeof(buf), "invalid response code %s (0x%04hx)", GetResponseName(r).c_str(), static_cast<unsigned short>(r));
 			return buf;
 		}
 	}
